pyproject.toml
[project]
name = "asia-in-brief"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.26.1",
    "feedparser>=6.0.12",
    "jinja2>=3.1.6",
    "openai>=2.9.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
]
